% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_alignment.R
\name{align_prcomp}
\alias{align_prcomp}
\title{Align an image and a mask based on principal components}
\usage{
align_prcomp(img = NULL, weighted = T, ...)
}
\arguments{
\item{img}{image or data frame of the locations of nonzero pixels in an image
(columns row, col, value*). If img is a data frame and has
column value, value will be used to weight the results. If img
is a list, each image in the list will be handled separately}

\item{weighted}{should weighted calculation be used?}

\item{...}{additional arguments to image_to_df}
}
\value{
angle to rotate the image by (in degrees)
}
\description{
One of img or img_df must be supplied. If img is supplied, additional
arguments to img_to_df may also be supplied using ....
}
