% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerate_img_control.R
\name{binary_center}
\alias{binary_center}
\title{Get the center of a binary image}
\usage{
binary_center(img, trim = T)
}
\arguments{
\item{img}{image/matrix}

\item{trim}{Trim 5\% from each side of the image? (Useful for removing page boundary issues)}
}
\description{
Calculate the center of mass of a binary image, with or without trimming
}
