% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerate_img_control.R
\name{clean_initial_img}
\alias{clean_initial_img}
\title{Invert, smooth, and threshold an image}
\usage{
clean_initial_img(img, gaussian_d = 25, threshold_val = 0.15)
}
\arguments{
\item{img}{Image}

\item{gaussian_d}{diameter of brush to use for gaussian blur}

\item{threshold_val}{threshold value to use on normalized, inverted, blurred
image. If threshold is 0, it will be automatically determined using
a heuristic approach.}
}
\description{
Invert, smooth, and threshold an image
}
