% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerate_img_auto.R
\name{em_thresh}
\alias{em_thresh}
\title{Compute likely pixel category based on EM algorithm clustering of intensity}
\usage{
em_thresh(img, scale_factor = 10, N = ifelse(is.null(scale_factor),
  pmin(length(img), 30000), NULL), ngroups = 3, quiet = T, ...)
}
\arguments{
\item{img}{Image}

\item{scale_factor}{Alternative to N - scales image by a factor of scale_factor}

\item{N}{Number of points to sample from the image (speeds up computational time)}

\item{ngroups}{Number of clusters}

\item{quiet}{suppress output from normalmixEM using sink()?}

\item{...}{additional arguments to mixtools::normalmixEM}
}
\description{
EM algorithm is for normally distributed groups; this assumption is likely
not accurate, but it is fast and effective.
If both N and scale_factor are NULL, the full image will be used, which will be slow.
}
