% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoe_mask.R
\name{est_ppi_film}
\alias{est_ppi_film}
\title{Estimate PPI for film images}
\usage{
est_ppi_film(dim)
}
\arguments{
\item{dim}{Image dimensions}
}
\description{
The film used in the longitudinal study has dimensions 7"x13".
This function is a convenience function which takes the dimensions of the
image and estimates the PPI so that the mask can be scaled appropriately.
}
