% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerate_img_auto.R
\name{exaggerate_img_auto}
\alias{exaggerate_img_auto}
\title{Exaggerate an image to a mask-like appearance automatically}
\usage{
exaggerate_img_auto(img, ...)
}
\arguments{
\item{img}{Image}

\item{...}{extra arguments, primarily to em_thresh}
}
\description{
Exaggerate an image to a mask-like appearance automatically
}
