% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerate_img_control.R
\name{exaggerate_img_control}
\alias{exaggerate_img_control}
\title{Exaggerate an image to a mask-like appearance with control of parameters}
\usage{
exaggerate_img_control(img, gaussian_d = 125, threshold_val = 0.125,
  opening_d = 7, closing_d = 301)
}
\arguments{
\item{img}{Image}

\item{gaussian_d}{diameter of brush to use for gaussian blur}

\item{threshold_val}{threshold value to use on normalized, inverted, blurred
image}

\item{opening_d}{diameter to use for image opening (despeckling)}

\item{closing_d}{diameter to use for image closing (exaggeration)}
}
\description{
Exaggerate an image to a mask-like appearance with control of parameters
}
