% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_alignment.R
\name{get_mask_arch}
\alias{get_mask_arch}
\title{Get arch location (approx) in a shoe mask}
\usage{
get_mask_arch(img)
}
\arguments{
\item{img}{Image}
}
\value{
center (row, col) of the location of the arch in the mask
}
\description{
This function assumes that the shoe is oriented such that the toe-heel axis
is vertical. The arch would then be the local minimum of mask width near the
center of the image.
}
\examples{
library(ShoeScrubR)
shoe_mask("Nike", 10, "L") \%>\%
get_mask_arch()
}
