% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_to_df}
\alias{image_to_df}
\title{Image to data frame}
\usage{
image_to_df(img, filter_val = NULL, row_neg = F)
}
\arguments{
\item{img}{image}

\item{filter_val}{value(s) of pixels to remove}

\item{row_neg}{negate the rows so that the image is "right side up"?}
}
\description{
Image to data frame
}
