% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_crop}
\alias{img_crop}
\title{Crop an image (and record metadata)}
\usage{
img_crop(img, dim, center = NULL)
}
\arguments{
\item{img}{Image or list of images}

\item{dim}{New dimension of the image}

\item{center}{point to use around which cropping is symmetrical}
}
\description{
Crop an image (and record metadata)
}
