% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_mode}
\alias{img_mode}
\title{Get most common pixel value (approx)}
\usage{
img_mode(img, digits = 2, size = 50000)
}
\arguments{
\item{img}{Image}

\item{digits}{precision of numerical values}

\item{size}{size of sample - if img has more than size pixels, it will be
reduced to a random sample of size before tabulation}
}
\description{
Get most common pixel value (approx)
}
