% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_pad}
\alias{img_pad}
\title{Pad an image}
\usage{
img_pad(img, padding = c(0, 0, 0, 0), value = 0, top = padding[1],
  bottom = padding[2], left = padding[3], right = padding[4])
}
\arguments{
\item{img}{image}

\item{padding}{vector of four padding values (top, bottom, left, right).
Specifying any one of the individual values will override the
vector specification.}

\item{value}{fill value to use for padding the image.
If value has the same length as the number of frames in the
image, it will be applied frame-wise.}

\item{top}{number of pixels to add to the top}

\item{bottom}{number of pixels to add to the bottom}

\item{left}{number of pixels to add to the left}

\item{right}{number of pixels to add to the right}
}
\description{
Pad an image
}
\examples{
par(mfrow = c(1, 2))
im <- EBImage::readImage(system.file('images', 'nuclei.tif', package='EBImage'))[,,1:3]
EBImage::colorMode(im) <- "Color"
dim(im)
plot(im)

im_pad <- img_pad(im, value = c(0, 1, 0), top = 15, bottom = 10, left = 5, right = 0)
dim(im_pad)
plot(im_pad)
im_pad <- img_pad(im, value = c(0, 1, 0), padding = c(15, 10, 5, 0))
dim(im_pad)
plot(im_pad)
}
