% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_pad_to_center}
\alias{img_pad_to_center}
\title{Pad image so that specified coordinates are in the center of the image}
\usage{
img_pad_to_center(img, center = round(dim(img)/2), value = 0)
}
\arguments{
\item{img}{image}

\item{center}{coordinates describing the desired "center" of the image (row, col)}

\item{value}{fill value to use for padding the image.
If value has the same length as the number of frames in the
image, it will be applied frame-wise.}
}
\description{
Pad image so that specified coordinates are in the center of the image
}
\examples{
par(mfrow = c(1, 2))
im <- EBImage::readImage(system.file('images', 'nuclei.tif', package='EBImage'))[,,1:3]
EBImage::colorMode(im) <- "Color"
dim(im)
plot(im)

im_pad <- img_pad_to_center(im, center = c(200, 200), value = c(0, 1, 0))
dim(im_pad)
plot(im_pad, all = TRUE)
}
