% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_pad_to_size}
\alias{img_pad_to_size}
\title{Pad image so that it is of the specified size}
\usage{
img_pad_to_size(img, size = dim(img), value = 0)
}
\arguments{
\item{img}{image}

\item{size}{dimensions of the output image (cannot be smaller than the
current dimensions)}

\item{value}{fill value to use for padding the image.
If value has the same length as the number of frames in the
image, it will be applied frame-wise.}
}
\description{
Pad image so that it is of the specified size
}
\examples{
par(mfrow = c(1, 2))
im <- EBImage::readImage(system.file('images', 'nuclei.tif', package='EBImage'))[,,1:3]
EBImage::colorMode(im) <- "Color"
dim(im)
plot(im)

im_pad <- img_pad_to_size(im, size = c(550, 540), value = c(0, 1, 0))
dim(im_pad)
plot(im_pad, all = TRUE)
}
