% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_pyramid}
\alias{img_pyramid}
\title{Image pyramid}
\usage{
img_pyramid(img, scale)
}
\arguments{
\item{img}{image (or list of images). If image list is named, resulting
tibble will have an extra column, img_name.}

\item{scale}{vector of numeric scaling factors}
}
\value{
tibble containing columns img, scale, dim, and (if original image
        list is named) img_name. The img column will contain the scaled image
}
\description{
Image pyramid
}
