% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_resize}
\alias{img_resize}
\title{Resize an image (and record metadata)}
\usage{
img_resize(img, ...)
}
\arguments{
\item{img}{Image}

\item{...}{additional arguments to EBImage::resize: width w, height h,
output.dim = c(w, h), output.origin = c(0, 0), antialias = F, ...
(other arguments to affine, including bg.col, antialias, filter)}
}
\value{
resized image with attribute "operation" that records the original
        and final dimensions of the image
}
\description{
Resize an image (and record metadata)
}
