% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_rotate}
\alias{img_rotate}
\title{Rotate an image (and record metadata)}
\usage{
img_rotate(img, angle, ...)
}
\arguments{
\item{img}{Image}

\item{angle}{rotation angle in degrees}

\item{...}{additional arguments to EBImage::rotate: output.dim,
output.origin, ... (other arguments to affine, including bg.col,
antialias, filter)}
}
\value{
resized image with attribute "operation" that records the original
        and final dimensions of the image
}
\description{
Rotate an image (and record metadata)
}
