% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_translate}
\alias{img_translate}
\title{Translate an image (and record metadata)}
\usage{
img_translate(img, v, ...)
}
\arguments{
\item{img}{Image}

\item{v}{translation vector (2 numbers)}

\item{...}{additional arguments to EBImage::translate: vector v, filter, ...
(other arguments to affine, including bg.col, antialias, filter)}
}
\value{
resized image with attribute "operation" that records the original
        and final dimensions of the image
}
\description{
Translate an image (and record metadata)
}
