% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_alignment.R
\name{rough_align}
\alias{rough_align}
\title{Initial alignment of mask and image}
\usage{
rough_align(img, mask, img_fill_value = img_mode(img), ...)
}
\arguments{
\item{img}{Image - greyscale, white(ish) as background, black as signal}

\item{mask}{Mask; white (signal) and black (bkgd)}

\item{img_fill_value}{value to fill any image padding with; defaults to the
mode value (as calculated by \code{\link{img_mode}}).}

\item{...}{extra parameters to pass to em_thresh for automatic thresholding}
}
\value{
list containing image and mask which have been aligned to center
        value and rotated via principal components so that PC1 is the
        positive y-axis.
}
\description{
Initial alignment of mask and image
}
