% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoe_mask.R
\name{shoe_mask}
\alias{shoe_mask}
\title{Get a specific shoe mask}
\usage{
shoe_mask(brand, size, foot, ppi = 300)
}
\arguments{
\item{brand}{Adidas or Nike}

\item{size}{Numerical size: 7, 7.5, 10, 10.5 for Adidas, 8, 8.5, 10, 10.5
for Nike}

\item{foot}{R or L, determined from the image (e.g. if the image is a film
print, this would be the opposite foot from the image name because
film prints are reversed)}

\item{ppi}{pixels per inch (linearly) for the returned mask image}
}
\description{
Get a specific shoe mask
}
\examples{
shoe_mask("Nike", 10.5, "R") \%>\% plot()
}
