% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShoeScrubR-package.R
\docType{package}
\name{ShoeScrubR-package}
\alias{ShoeScrubR}
\alias{ShoeScrubR-package}
\title{ShoeScrubR: Functions for cleaning data from CSAFE's Longitudinal Shoe Study}
\description{
The CSAFE Longitudinal Shoe Study collected data on 160 pairs of 
    shoes over 4 observation points approximately 5 weeks apart. Additional data
    such as number of steps, participant weight, and activities was also 
    collected as part of the study. This package contains some utilities for 
    cleaning the original data up for statistical analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/srvanderplas/ShoeScrubR}
  \item Report bugs at \url{https://github.com/srvanderplas/ShoeScrubR/issues}
}

}
\author{
\strong{Maintainer}: Susan Vanderplas \email{srvanderplas@gmail.com}

}
\keyword{internal}
